requirements.txt
Django==1.11.4
Pillow==4.2.1
djangorestframework==3.6.3
olefile==0.44
pytz==2017.2
